Fix issues with cassandra querying on role_assignment table
We cannot query based on all 4 columns of the table, but only on the columns
which form the primary key. This means we need to query several times, or
get large data and filter in-memory for our desired row -- a sub-optimal code.

Must be looked at in future, for performance implications. Looks fine for now
as it just increases number of queries 4 times at max, which is better than
O(n) times.

+++ keystone/assignment/backends/cass.py
@@ -324,13 +324,12 @@ class Assignment(keystone_assignment.Driver):
         # the case when the exception will be thrown. Another alternative would
         # be to just do a write, which will overwrite a previous value if it
         # existed, and this won't raise an exception
-        ref = RoleAssignment(
+        #import pdb; pdb.set_trace()
+        refs = RoleAssignment.objects.filter(
                 type=AssignmentType.USER_PROJECT,
-                actor_id=user_id,
-                target_id=tenant_id,
-                role_id=role_id,
-                inherited=False)
-        if len(ref) != 0:
+                actor_id=user_id)
+        refs = [ref for ref in refs if ref.target_id == tenant_id and ref.role_id == role_id ]
+        if len(refs) != 0:
             msg = ('User %s already has role %s in tenant %s'
                    % (user_id, role_id, tenant_id))
             raise exception.Conflict(type='role grant', details=msg)
@@ -343,13 +342,18 @@ class Assignment(keystone_assignment.Driver):
             inherited=False)
 
     def remove_role_from_user_and_project(self, user_id, tenant_id, role_id):
-        try:
-            ref = RoleAssignment.get(
-                actor_id=user_id,
-                target_id=tenant_id,
-                role_id=role_id)
-            ref.delete()
-        except DoesNotExist:
+        role_found = False
+        for type in [AssignmentType.USER_PROJECT,
+                AssignmentType.USER_DOMAIN,
+                AssignmentType.GROUP_PROJECT,
+                AssignmentType.GROUP_DOMAIN]:
+            refs = RoleAssignment.filter(type=type, actor_id=user_id)
+            for ref in refs:
+                if ref.target_id == tenant_id and ref.role_id == role_id:
+                    role_found = True
+                    ref.delete()
+
+        if not role_found:
             raise exception.RoleNotFound(message=_(
                 'Cannot remove role that has not been granted, %s') %
                 role_id)
